move button to home view
